Link internal C++ authentication plugins inside client library
