+++ pulsar-client-cpp/lib/auth/CMakeLists.txt
@@ -17,12 +17,7 @@
 # under the License.
 #
 
-add_library(authTls MODULE AuthTls.cc)
-target_link_libraries(authTls ${CLIENT_LIBS})
-set_target_properties(authTls PROPERTIES OUTPUT_NAME authtls)
-
-add_library(authAthenz MODULE AuthAthenz.cc)
-target_link_libraries(authAthenz ${CLIENT_LIBS} ztsClient)
-set_target_properties(authAthenz PROPERTIES OUTPUT_NAME authathenz)
+add_library(authTls STATIC AuthTls.cc)
+add_library(authAthenz STATIC AuthAthenz.cc)
 
 add_subdirectory(athenz)
